requirements.txt
groq
python-dotenv
flask
flask-cors
dotenv
requests
mysql-connector-python
